backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
celery==5.5.3
click==8.2.1
fastapi==0.116.1
google-api-python-client==2.179.0
h11==0.16.0
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
qdrant-client==1.15.1
redis==6.4.0
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
webvtt-py==0.5.1
pytest==8.4.1
youtube-transcript-api==1.2.2
yt-dlp==2024.12.13
llama-index-core==0.13.3
llama-index-embeddings-ollama==0.8.2
llama-index-embeddings-google-genai==0.3.0
llama-index-vector-stores-qdrant==0.8.2
langgraph==0.6.6
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
tavily-python==0.7.11
langchain-anthropic==0.3.19
langchain-google-genai==2.1.10
langgraph-checkpoint-sqlite==2.0.11
fastembed==0.7.3
llama-index-llms-openai==0.5.4
pyhumps==3.8.0
llama-index-embeddings-fastembed==0.4.0
rank-bm25==0.2.2
FlagEmbedding==1.3.5
watchdog==6.0.0
# PDF processing packages
PyMuPDF==1.24.14  # fitz for PDF text extraction
pdfplumber==0.11.4  # Alternative PDF processor
# Arxiv package
arxiv==2.1.3
# HTML scraping packages
requests-html==0.10.0
lxml==5.3.0
html5lib==1.1
# Additional scraping utilities
fake-useragent==1.5.1
urllib3==2.2.3
python-multipart==0.0.20
arize-phoenix==11.30.0
langgraph-checkpoint-postgres==2.0.23 
asyncpg==0.30.0
llama-index-llms-ollama==0.7.1
